pyproject.toml
[project]
name = "obsidian-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gkeepapi>=0.17.1",
    "google-api-python-client>=2.193.0",
    "google-auth-httplib2>=0.3.1",
    "google-auth-oauthlib>=1.3.1",
    "google-genai>=1.70.0",
    "gpsoauth>=2.0.0",
    "macnotesapp>=0.8.2",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
]
